+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## TBD
+## 4.21.0 (2019-10-14)
 
 * Collect additional data in internal error reports
   [#612](https://github.com/bugsnag/bugsnag-android/pull/612)